New: added cool new feature to parallelize the execution of a chain
In more detail: a new feature to parallelize (fork) the execution of (the links in) a chain.

For an example, see the macro esk112_parallel_fork_demo and the link ForkExample.
The link ForkDataCollector can collect and share data objects between different forked processes,
and can paste them together. It does do by making use of the ForkStore, which is a shared dict
between forked processed. The ForkStore is similar to the DataStore, but gets initialized and
cleared at the start and end of the (forked) chain respectively. It is not persisted.
When putting or reading objects from the ForkStore, for stability the ForkStore needs to be locked.
(See ForkExample and ForkDataCollector links.)
Note the the forking of the chain is only for the execution of links/chain, not initialize and finalize.
The number of forks running concurrently is by default set to the number of cpu cores, but
can be set by hand as well. (See example macro.)